Add pre and post transaction pacman hooks
These pacman hooks run whenever the system updates the boot images, systemd, or kernel microcode.

The pre-transaction hook to detects: 1) modifications to the boot files which the user has been notified of but has not yet acknowledged, and 2) modifications to the boot files which have taken place since the last boot and the user is unaware of.

This verifies that no unauthorized modifications have taken place while the system has been in use.  It is necessary since the post-transaction hook assumes good files and updates all hashes accordingly.  A malicious bootkit that is aware of chkboot's existence on the system can, of course, simply update the hashes itself to bypass this verification.

The post-transaction hook assumes good files and updates all hashes accordingly.

These hooks require pull requests #9 and #10 to function properly

Additionally requires pull request #11 for systems using a vfat boot partition